tutorial_restfulapi/requirements.txt
Django==2.0.3
django-rest-framework==0.1.0
djangorestframework==3.7.7
pytz==2018.3
drf-nested-routers==0.90.0
